GPLIM-5815 Code review changes.  Changed method to create a test SAP quote to accept doubles instead of Big Decimal.

+++ mercury/src/test/java/org/broadinstitute/gpinformatics/athena/boundary/billing/BillingCreditDbFreeTest.java
@@ -205,7 +205,7 @@ public class BillingCreditDbFreeTest {
         final String firstDeliveryDocument = "0211403";
 
         if (quoteSourceType == ProductOrder.QuoteSourceType.SAP_SOURCE) {
-            SapQuote sapQuote = TestUtils.buildTestSapQuote(pdo.getQuoteId(), BigDecimal.valueOf(10000),BigDecimal.valueOf(100000),
+            SapQuote sapQuote = TestUtils.buildTestSapQuote(pdo.getQuoteId(), 10000d, 100000d,
                     pdo, TestUtils.SapQuoteTestScenario.DOLLAR_LIMITED, "GP01");
 
             Mockito.when(mockSapClient.findQuoteDetails(Mockito.anyString())).thenReturn(sapQuote);
@@ -294,7 +294,7 @@ public class BillingCreditDbFreeTest {
         final String firstDeliveryDocumentId = "0211403";
 
         if (quoteSourceType == ProductOrder.QuoteSourceType.SAP_SOURCE) {
-            SapQuote sapQuote = TestUtils.buildTestSapQuote(pdo.getQuoteId(), BigDecimal.valueOf(10000),BigDecimal.valueOf(100000),
+            SapQuote sapQuote = TestUtils.buildTestSapQuote(pdo.getQuoteId(), 10000d, 100000d,
                     pdo, TestUtils.SapQuoteTestScenario.DOLLAR_LIMITED, "GP01");
             Mockito.when(mockSapClient.findQuoteDetails(Mockito.anyString())).thenReturn(sapQuote);
             Mockito.when(mockSapClient.createDeliveryDocument(Mockito.any(SAPDeliveryDocument.class))).thenReturn(
@@ -394,7 +394,7 @@ public class BillingCreditDbFreeTest {
             pdo.setQuoteId("99339288");
         }
 
-        SapQuote sapQuote = TestUtils.buildTestSapQuote(pdo.getQuoteId(), BigDecimal.valueOf(10000),BigDecimal.valueOf(100000),
+        SapQuote sapQuote = TestUtils.buildTestSapQuote(pdo.getQuoteId(), 10000d, 100000d,
                 pdo, TestUtils.SapQuoteTestScenario.DOLLAR_LIMITED, "GP01");
 
         if (quoteSourceType == ProductOrder.QuoteSourceType.SAP_SOURCE) {
@@ -452,7 +452,7 @@ public class BillingCreditDbFreeTest {
         }
 
         if (quoteSourceType == ProductOrder.QuoteSourceType.SAP_SOURCE) {
-            SapQuote sapQuote = TestUtils.buildTestSapQuote(pdo.getQuoteId(), BigDecimal.valueOf(10000),BigDecimal.valueOf(100000),
+            SapQuote sapQuote = TestUtils.buildTestSapQuote(pdo.getQuoteId(), 10000d, 100000d,
                     pdo, TestUtils.SapQuoteTestScenario.DOLLAR_LIMITED, "GP01");
 
             Mockito.when(mockSapClient.findQuoteDetails(Mockito.anyString())).thenReturn(sapQuote);
@@ -516,7 +516,7 @@ public class BillingCreditDbFreeTest {
         }
 
         if (quoteSourceType == ProductOrder.QuoteSourceType.SAP_SOURCE) {
-            SapQuote sapQuote = TestUtils.buildTestSapQuote(pdo.getQuoteId(), BigDecimal.valueOf(10000),BigDecimal.valueOf(100000),
+            SapQuote sapQuote = TestUtils.buildTestSapQuote(pdo.getQuoteId(), 10000d, 100000d,
                     pdo, TestUtils.SapQuoteTestScenario.DOLLAR_LIMITED, "GP01");
 
             Mockito.when(mockSapClient.findQuoteDetails(Mockito.anyString())).thenReturn(sapQuote);
